gRPC controller
