--- telemetry/collector/src/main/java/collector/Collector.java
@@ -1,13 +1,0 @@
-package collector;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-
-@SpringBootApplication
-@ConfigurationPropertiesScan
-public class Collector {
-    public static void main(String[] args) {
-        SpringApplication.run(Collector.class, args);
-    }
-}
