--- telemetry/collector/src/main/java/collector/controller/ErrorMessage.java
@@ -1,13 +1,0 @@
-package collector.controller;
-
-public class ErrorMessage {
-    private final String error;
-
-    public ErrorMessage(String error) {
-        this.error = error;
-    }
-
-    public String getError() {
-        return error;
-    }
-}
