--- telemetry/collector/src/main/java/collector/controller/EventController.java
@@ -1,70 +1,0 @@
-package collector.controller;
-
-import collector.handler.hub.HubEventHandler;
-import collector.handler.sensor.SensorEventHandler;
-import collector.model.hub.HubEvent;
-import collector.model.hub.HubEventType;
-import collector.model.sensor.SensorEvent;
-import collector.model.sensor.SensorEventType;
-import jakarta.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-@RestController
-@RequestMapping("/events")
-public class EventController {
-    private final Map<SensorEventType, SensorEventHandler> sensorEventHandlers;
-    private final Map<HubEventType, HubEventHandler> hubEventHandlers;
-
-    public EventController(List<SensorEventHandler> sensorEventHandlers,
-                           List<HubEventHandler> hubEventHandlers) {
-        this.sensorEventHandlers = sensorEventHandlers.stream()
-                .collect(Collectors.toMap(SensorEventHandler::getEventType, Function.identity()));
-
-        this.hubEventHandlers = hubEventHandlers.stream()
-                .collect(Collectors.toMap(HubEventHandler::getEventType, Function.identity()));
-    }
-
-    @PostMapping("/sensors")
-    @ResponseStatus(HttpStatus.OK)
-    public void collectSensorEvent(@Valid @RequestBody SensorEvent request) {
-        if (sensorEventHandlers.containsKey(request.getType())) {
-            sensorEventHandlers.get(request.getType()).handle(request);
-        } else {
-            throw new IllegalArgumentException(String.format("Для события сенсора %s не существует обработчика",
-                    request.getType()));
-        }
-    }
-
-    @PostMapping("/hubs")
-    @ResponseStatus(HttpStatus.OK)
-    public void collectHubEvent(@Valid @RequestBody HubEvent request) {
-        if (hubEventHandlers.containsKey(request.getType())) {
-            hubEventHandlers.get(request.getType()).handle(request);
-        } else {
-            throw new IllegalArgumentException(String.format("Для события хаба %s не существует обработчика",
-                    request.getType()));
-        }
-    }
-
-    // обработка ошибок валидации @Valid
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorMessage> handleException(MethodArgumentNotValidException exception) {
-        String message = exception.getFieldErrors().stream()
-                .map(fieldError ->
-                        String.format("Поле %s %s.", fieldError.getField(), fieldError.getDefaultMessage()))
-                .reduce(" ", String::concat);
-        message = message.strip();
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(new ErrorMessage(message));
-    }
-}
