--- telemetry/collector/src/main/java/collector/handler/hub/DeviceRemovedEventHandler.java
@@ -1,33 +1,0 @@
-package collector.handler.hub;
-
-import collector.kafka.Producer;
-import collector.model.hub.DeviceRemovedEvent;
-import collector.model.hub.HubEvent;
-import collector.model.hub.HubEventType;
-import org.springframework.stereotype.Component;
-import ru.yandex.practicum.kafka.telemetry.event.DeviceRemovedEventAvro;
-
-/**
- * Обработчик события удаления устройства из хаба
- */
-@Component
-public class DeviceRemovedEventHandler extends BaseHubEventHandler<DeviceRemovedEventAvro> {
-
-    public DeviceRemovedEventHandler(Producer producer) {
-        super(producer);
-    }
-
-    @Override
-    protected DeviceRemovedEventAvro mapToAvro(HubEvent event) {
-        DeviceRemovedEvent deviceRemovedEvent = (DeviceRemovedEvent) event;
-
-        return DeviceRemovedEventAvro.newBuilder()
-                .setId(deviceRemovedEvent.getId())
-                .build();
-    }
-
-    @Override
-    public HubEventType getEventType() {
-        return HubEventType.DEVICE_REMOVED_EVENT;
-    }
-}
