--- telemetry/collector/src/main/java/collector/handler/hub/HubEventHandler.java
@@ -1,13 +1,0 @@
-package collector.handler.hub;
-
-import collector.model.hub.HubEvent;
-import collector.model.hub.HubEventType;
-
-/**
- * Интерфейс, объединяющий все обработчики для событий HubEvent
- */
-public interface HubEventHandler {
-    void handle(HubEvent event);
-
-    HubEventType getEventType();
-}
