--- telemetry/collector/src/main/java/collector/handler/sensor/ClimateSensorEventHandler.java
@@ -1,35 +1,0 @@
-package collector.handler.sensor;
-
-import collector.kafka.Producer;
-import collector.model.sensor.ClimateSensorEvent;
-import collector.model.sensor.SensorEvent;
-import collector.model.sensor.SensorEventType;
-import org.springframework.stereotype.Component;
-import ru.yandex.practicum.kafka.telemetry.event.ClimateSensorAvro;
-
-/**
- * Обработчик события от климатического датчика
- */
-@Component
-public class ClimateSensorEventHandler extends BaseSensorEventHandler<ClimateSensorAvro> {
-
-    public ClimateSensorEventHandler(Producer producer) {
-        super(producer);
-    }
-
-    @Override
-    protected ClimateSensorAvro mapToAvro(SensorEvent event) {
-        ClimateSensorEvent climateSensorEvent = (ClimateSensorEvent) event;
-
-        return ClimateSensorAvro.newBuilder()
-                .setCo2Level(climateSensorEvent.getCo2Level())
-                .setHumidity(climateSensorEvent.getHumidity())
-                .setTemperatureC(climateSensorEvent.getTemperatureC())
-                .build();
-    }
-
-    @Override
-    public SensorEventType getEventType() {
-        return SensorEventType.CLIMATE_SENSOR_EVENT;
-    }
-}
