--- telemetry/collector/src/main/java/collector/handler/sensor/MotionSensorEventHandler.java
@@ -1,34 +1,0 @@
-package collector.handler.sensor;
-
-import collector.kafka.Producer;
-import collector.model.sensor.MotionSensorEvent;
-import collector.model.sensor.SensorEvent;
-import collector.model.sensor.SensorEventType;
-import org.springframework.stereotype.Component;
-import ru.yandex.practicum.kafka.telemetry.event.MotionSensorAvro;
-
-/**
- * Обработчик события от датчика движения
- */
-@Component
-public class MotionSensorEventHandler extends BaseSensorEventHandler<MotionSensorAvro> {
-    public MotionSensorEventHandler(Producer producer) {
-        super(producer);
-    }
-
-    @Override
-    protected MotionSensorAvro mapToAvro(SensorEvent event) {
-        MotionSensorEvent motionSensorEvent = (MotionSensorEvent) event;
-
-        return MotionSensorAvro.newBuilder()
-                .setMotion(motionSensorEvent.getMotion())
-                .setLinkQuality(motionSensorEvent.getLinkQuality())
-                .setVoltage(motionSensorEvent.getVoltage())
-                .build();
-    }
-
-    @Override
-    public SensorEventType getEventType() {
-        return SensorEventType.MOTION_SENSOR_EVENT;
-    }
-}
