--- telemetry/collector/src/main/java/collector/handler/sensor/SensorEventHandler.java
@@ -1,13 +1,0 @@
-package collector.handler.sensor;
-
-import collector.model.sensor.SensorEvent;
-import collector.model.sensor.SensorEventType;
-
-/**
- * Интерфейс, объединяющий все обработчики для событий SensorEvent
- */
-public interface SensorEventHandler {
-    void handle(SensorEvent event);
-
-    SensorEventType getEventType();
-}
