--- telemetry/collector/src/main/java/collector/handler/sensor/SwitchSensorEventHandler.java
@@ -1,32 +1,0 @@
-package collector.handler.sensor;
-
-import collector.kafka.Producer;
-import collector.model.sensor.SensorEvent;
-import collector.model.sensor.SensorEventType;
-import collector.model.sensor.SwitchSensorEvent;
-import org.springframework.stereotype.Component;
-import ru.yandex.practicum.kafka.telemetry.event.SwitchSensorAvro;
-
-/**
- * Обработчик события от датчика переключателя
- */
-@Component
-public class SwitchSensorEventHandler extends BaseSensorEventHandler<SwitchSensorAvro> {
-    public SwitchSensorEventHandler(Producer producer) {
-        super(producer);
-    }
-
-    @Override
-    protected SwitchSensorAvro mapToAvro(SensorEvent event) {
-        SwitchSensorEvent switchSensorEvent = (SwitchSensorEvent) event;
-
-        return SwitchSensorAvro.newBuilder()
-                .setState(switchSensorEvent.getState())
-                .build();
-    }
-
-    @Override
-    public SensorEventType getEventType() {
-        return SensorEventType.SWITCH_SENSOR_EVENT;
-    }
-}
