--- telemetry/collector/src/main/java/collector/handler/sensor/TemperatureSensorEventHandler.java
@@ -1,33 +1,0 @@
-package collector.handler.sensor;
-
-import collector.kafka.Producer;
-import collector.model.sensor.SensorEvent;
-import collector.model.sensor.SensorEventType;
-import collector.model.sensor.TemperatureSensorEvent;
-import org.springframework.stereotype.Component;
-import ru.yandex.practicum.kafka.telemetry.event.TemperatureSensorAvro;
-
-/**
- * Обработчик события от температурного датчика
- */
-@Component
-public class TemperatureSensorEventHandler extends BaseSensorEventHandler<TemperatureSensorAvro> {
-    public TemperatureSensorEventHandler(Producer producer) {
-        super(producer);
-    }
-
-    @Override
-    protected TemperatureSensorAvro mapToAvro(SensorEvent event) {
-        TemperatureSensorEvent temperatureSensorEvent = (TemperatureSensorEvent) event;
-
-        return TemperatureSensorAvro.newBuilder()
-                .setTemperatureC(temperatureSensorEvent.getTemperatureC())
-                .setTemperatureF(temperatureSensorEvent.getTemperatureF())
-                .build();
-    }
-
-    @Override
-    public SensorEventType getEventType() {
-        return SensorEventType.TEMPERATURE_SENSOR_EVENT;
-    }
-}
