--- telemetry/collector/src/main/java/collector/kafka/TopicType.java
@@ -1,6 +1,0 @@
-package collector.kafka;
-
-public enum TopicType {
-    SENSORS_EVENTS_TOPIC,
-    HUBS_EVENTS_TOPIC
-}
