--- telemetry/collector/src/main/java/collector/model/hub/DeviceRemovedEvent.java
@@ -1,28 +1,0 @@
-package collector.model.hub;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.experimental.FieldDefaults;
-
-/**
- * Событие, сигнализирующее об удалении устройства из системы/
- */
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Getter
-@Setter
-@ToString(callSuper = true)
-public class DeviceRemovedEvent extends HubEvent {
-    /**
-     * Идентификатор удаленного устройства.
-     */
-    @NotNull
-    String id;
-
-    @Override
-    public HubEventType getType() {
-        return HubEventType.DEVICE_REMOVED_EVENT;
-    }
-}
