--- telemetry/collector/src/main/java/collector/model/hub/DeviceType.java
@@ -1,12 +1,0 @@
-package collector.model.hub;
-
-/**
- * Перечисление типов устройств, которые могут быть добавлены в систему.
- */
-public enum DeviceType {
-    MOTION_SENSOR,
-    TEMPERATURE_SENSOR,
-    LIGHT_SENSOR,
-    CLIMATE_SENSOR,
-    SWITCH_SENSOR
-}
