--- telemetry/collector/src/main/java/collector/model/hub/HubEvent.java
@@ -1,47 +1,0 @@
-package collector.model.hub;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.experimental.FieldDefaults;
-
-import java.time.Instant;
-
-/**
- * Базовый класс события хаба.
- */
-@JsonTypeInfo(
-        use = JsonTypeInfo.Id.NAME,
-        include = JsonTypeInfo.As.EXISTING_PROPERTY,
-        property = "type"
-)
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = DeviceAddedEvent.class, name = "DEVICE_ADDED"),
-        @JsonSubTypes.Type(value = DeviceRemovedEvent.class, name = "DEVICE_REMOVED"),
-        @JsonSubTypes.Type(value = ScenarioAddedEvent.class, name = "SCENARIO_ADDED"),
-        @JsonSubTypes.Type(value = ScenarioRemovedEvent.class, name = "SCENARIO_REMOVED")
-})
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
-@Data
-public abstract class HubEvent {
-    /**
-     * Идентификатор хаба, связанный с событием.
-     */
-    @NotNull
-    @EqualsAndHashCode.Include
-    String hubId;
-
-    /**
-     * Временная метка события. По умолчанию устанавливается текущее время.
-     */
-    Instant timestamp = Instant.now();
-
-    /**
-     * Тип события хаба.
-     */
-    public abstract HubEventType getType();
-}
