--- telemetry/collector/src/main/java/collector/model/hub/HubEventType.java
@@ -1,11 +1,0 @@
-package collector.model.hub;
-
-/**
- * Перечисление возможных типов событий хаба.
- */
-public enum HubEventType {
-    DEVICE_ADDED_EVENT,
-    DEVICE_REMOVED_EVENT,
-    SCENARIO_ADDED_EVENT,
-    SCENARIO_REMOVED_EVENT
-}
