--- telemetry/collector/src/main/java/collector/model/sensor/ClimateSensorEvent.java
@@ -1,26 +1,0 @@
-package collector.model.sensor;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@ToString(callSuper = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class ClimateSensorEvent extends SensorEvent {
-    @NotNull
-    Integer temperatureC;
-
-    @NotNull
-    Integer humidity;
-
-    @NotNull
-    Integer co2Level;
-
-    @Override
-    public SensorEventType getType() {
-        return SensorEventType.CLIMATE_SENSOR_EVENT;
-    }
-}
-
