--- telemetry/collector/src/main/java/collector/model/sensor/LightSensorEvent.java
@@ -1,25 +1,0 @@
-package collector.model.sensor;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@ToString(callSuper = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class LightSensorEvent extends SensorEvent {
-    @NotNull
-    Integer linkQuality;
-
-    @NotNull
-    Integer luminosity;
-
-    @Override
-    public SensorEventType getType() {
-        return SensorEventType.LIGHT_SENSOR_EVENT;
-    }
-}
