--- telemetry/collector/src/main/java/collector/model/sensor/MotionSensorEvent.java
@@ -1,28 +1,0 @@
-package collector.model.sensor;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@ToString(callSuper = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class MotionSensorEvent extends SensorEvent {
-    @NotNull
-    Integer linkQuality;
-
-    @NotNull
-    Boolean motion;
-
-    @NotNull
-    Integer voltage;
-
-    @Override
-    public SensorEventType getType() {
-        return SensorEventType.MOTION_SENSOR_EVENT;
-    }
-}
