--- telemetry/collector/src/main/java/collector/model/sensor/SensorEventType.java
@@ -1,9 +1,0 @@
-package collector.model.sensor;
-
-public enum SensorEventType {
-    CLIMATE_SENSOR_EVENT,
-    LIGHT_SENSOR_EVENT,
-    MOTION_SENSOR_EVENT,
-    SWITCH_SENSOR_EVENT,
-    TEMPERATURE_SENSOR_EVENT
-}
