--- telemetry/collector/src/main/java/collector/model/sensor/SwitchSensorEvent.java
@@ -1,22 +1,0 @@
-package collector.model.sensor;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@ToString(callSuper = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class SwitchSensorEvent extends SensorEvent {
-    @NotNull
-    Boolean state;
-
-    @Override
-    public SensorEventType getType() {
-        return SensorEventType.SWITCH_SENSOR_EVENT;
-    }
-}
