--- telemetry/collector/src/main/java/collector/model/sensor/TemperatureSensorEvent.java
@@ -1,25 +1,0 @@
-package collector.model.sensor;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@ToString(callSuper = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class TemperatureSensorEvent extends SensorEvent {
-    @NotNull
-    Integer temperatureC;
-
-    @NotNull
-    Integer temperatureF;
-
-    @Override
-    public SensorEventType getType() {
-        return SensorEventType.TEMPERATURE_SENSOR_EVENT;
-    }
-}
