+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/Collector.java
@@ -1,0 +1,13 @@
+package ru.yandex.practicum.telemetry.collector;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+
+@SpringBootApplication
+@ConfigurationPropertiesScan
+public class Collector {
+    public static void main(String[] args) {
+        SpringApplication.run(Collector.class, args);
+    }
+}
