+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/controller/EventController.java
@@ -1,0 +1,76 @@
+package ru.yandex.practicum.telemetry.collector.controller;
+
+import com.google.protobuf.Empty;
+import io.grpc.Status;
+import io.grpc.StatusRuntimeException;
+import io.grpc.stub.StreamObserver;
+import lombok.extern.slf4j.Slf4j;
+import net.devh.boot.grpc.server.service.GrpcService;
+import ru.yandex.practicum.grpc.telemetry.collector.CollectorControllerGrpc.CollectorControllerImplBase;
+import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.telemetry.collector.handler.hub.HubEventHandler;
+import ru.yandex.practicum.telemetry.collector.handler.sensor.SensorEventHandler;
+
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+@GrpcService
+@Slf4j
+public class EventController extends CollectorControllerImplBase {
+    private final Map<SensorEventProto.PayloadCase, SensorEventHandler> sensorEventHandlers;
+    private final Map<HubEventProto.PayloadCase, HubEventHandler> hubEventHandlers;
+
+    public EventController(List<SensorEventHandler> sensorEventHandlers,
+                           List<HubEventHandler> hubEventHandlers) {
+        this.sensorEventHandlers = sensorEventHandlers.stream()
+                .collect(Collectors.toMap(SensorEventHandler::getEventType, Function.identity()));
+
+        this.hubEventHandlers = hubEventHandlers.stream()
+                .collect(Collectors.toMap(HubEventHandler::getEventType, Function.identity()));
+    }
+
+    @Override
+    public void collectSensorEvent(SensorEventProto request, StreamObserver<Empty> responseObserver) {
+        log.debug("Collect sensor event, request:\n{}", request.toString());
+        try {
+            // проверяем, есть ли обработчик для полученного события
+            if (sensorEventHandlers.containsKey(request.getPayloadCase())) {
+                // если обработчик найден, передаём событие ему на обработку
+                sensorEventHandlers.get(request.getPayloadCase()).handle(request);
+            } else {
+                throw new IllegalArgumentException("Not found handler for sensor event type: " + request.getPayloadCase());
+            }
+
+            // после обработки события возвращаем ответ клиенту
+            responseObserver.onNext(Empty.getDefaultInstance());
+            // и завершаем обработку запроса
+            responseObserver.onCompleted();
+        } catch (Exception e) {
+            // в случае исключения отправляем ошибку клиенту
+            log.error(e.getMessage());
+            responseObserver.onError(new StatusRuntimeException(Status.fromThrowable(e)));
+        }
+    }
+
+
+    @Override
+    public void collectHubEvent(HubEventProto request, StreamObserver<Empty> responseObserver) {
+        log.debug("Collect hub event, request:\n{}", request.toString());
+        try {
+            if (hubEventHandlers.containsKey(request.getPayloadCase())) {
+                hubEventHandlers.get(request.getPayloadCase()).handle(request);
+            } else {
+                throw new IllegalArgumentException("Not found handler for hub event type: " + request.getPayloadCase());
+            }
+
+            responseObserver.onNext(Empty.getDefaultInstance());
+            responseObserver.onCompleted();
+        } catch (Exception e) {
+            log.error(e.getMessage());
+            responseObserver.onError(new StatusRuntimeException(Status.fromThrowable(e)));
+        }
+    }
+}
