+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/hub/DeviceAddedEventHandler.java
@@ -1,0 +1,34 @@
+package ru.yandex.practicum.telemetry.collector.handler.hub;
+
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.DeviceAddedEventProto;
+import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.DeviceAddedEventAvro;
+import ru.yandex.practicum.kafka.telemetry.event.DeviceTypeAvro;
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+
+/**
+ * Обработчик события добавления устройства в хаб
+ */
+@Component
+public class DeviceAddedEventHandler extends BaseHubEventHandler<DeviceAddedEventAvro> {
+
+    public DeviceAddedEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    public HubEventProto.PayloadCase getEventType() {
+        return HubEventProto.PayloadCase.DEVICE_ADDED;
+    }
+
+    @Override
+    protected DeviceAddedEventAvro mapToAvro(HubEventProto event) {
+        DeviceAddedEventProto deviceAddedEvent = event.getDeviceAdded();
+
+        return DeviceAddedEventAvro.newBuilder()
+                .setId(deviceAddedEvent.getId())
+                .setType(DeviceTypeAvro.valueOf(deviceAddedEvent.getType().name()))
+                .build();
+    }
+}
