+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/hub/DeviceRemovedEventHandler.java
@@ -1,0 +1,32 @@
+package ru.yandex.practicum.telemetry.collector.handler.hub;
+
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.DeviceRemovedEventProto;
+import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.DeviceRemovedEventAvro;
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+
+/**
+ * Обработчик события удаления устройства из хаба
+ */
+@Component
+public class DeviceRemovedEventHandler extends BaseHubEventHandler<DeviceRemovedEventAvro> {
+
+    public DeviceRemovedEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    protected DeviceRemovedEventAvro mapToAvro(HubEventProto event) {
+        DeviceRemovedEventProto deviceRemovedEvent = event.getDeviceRemoved();
+
+        return DeviceRemovedEventAvro.newBuilder()
+                .setId(deviceRemovedEvent.getId())
+                .build();
+    }
+
+    @Override
+    public HubEventProto.PayloadCase getEventType() {
+        return HubEventProto.PayloadCase.DEVICE_REMOVED;
+    }
+}
