+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/hub/ScenarioAddedEventHandler.java
@@ -1,0 +1,55 @@
+package ru.yandex.practicum.telemetry.collector.handler.hub;
+
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
+import ru.yandex.practicum.grpc.telemetry.event.ScenarioAddedEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.*;
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+
+import java.util.List;
+
+/**
+ * Обработчик события добавления сценария в хаб
+ */
+@Component
+public class ScenarioAddedEventHandler extends BaseHubEventHandler<ScenarioAddedEventAvro> {
+
+    public ScenarioAddedEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    protected ScenarioAddedEventAvro mapToAvro(HubEventProto event) {
+        ScenarioAddedEventProto scenarioAddedEvent = event.getScenarioAdded();
+
+        List<DeviceActionAvro> actions = scenarioAddedEvent.getScenarioActionsList().stream()
+                .map(action ->
+                        DeviceActionAvro.newBuilder()
+                        .setSensorId(action.getSensorId())
+                        .setType(ActionTypeAvro.valueOf(action.getType().name()))
+                        .setValue(action.getValue())
+                        .build())
+                .toList();
+
+        List<ScenarioConditionAvro> scenarioConditions = scenarioAddedEvent.getScenarioConditionsList().stream()
+                .map(condition ->
+                        ScenarioConditionAvro.newBuilder()
+                                .setSensorId(condition.getSensorId())
+                                .setType(ConditionTypeAvro.valueOf(condition.getType().name()))
+                                .setOperation(ConditionOperationAvro.valueOf(condition.getOperation().name()))
+                                .setValue(condition.hasBoolValue() ? condition.getBoolValue() : condition.getIntValue())
+                                .build())
+                .toList();
+
+        return ScenarioAddedEventAvro.newBuilder()
+                .setName(scenarioAddedEvent.getName())
+                .setActions(actions)
+                .setConditions(scenarioConditions)
+                .build();
+    }
+
+    @Override
+    public HubEventProto.PayloadCase getEventType() {
+        return HubEventProto.PayloadCase.SCENARIO_ADDED;
+    }
+}
