+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/sensor/ClimateSensorEventHandler.java
@@ -1,0 +1,34 @@
+package ru.yandex.practicum.telemetry.collector.handler.sensor;
+
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.ClimateSensorProto;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.ClimateSensorAvro;
+
+/**
+ * Обработчик события от климатического датчика
+ */
+@Component
+public class ClimateSensorEventHandler extends BaseSensorEventHandler<ClimateSensorAvro> {
+
+    public ClimateSensorEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    protected ClimateSensorAvro mapToAvro(SensorEventProto event) {
+        ClimateSensorProto climateSensorEvent = event.getClimateSensorEvent();
+
+        return ClimateSensorAvro.newBuilder()
+                .setCo2Level(climateSensorEvent.getCo2Level())
+                .setHumidity(climateSensorEvent.getHumidity())
+                .setTemperatureC(climateSensorEvent.getTemperatureC())
+                .build();
+    }
+
+    @Override
+    public SensorEventProto.PayloadCase getEventType() {
+        return SensorEventProto.PayloadCase.CLIMATE_SENSOR_EVENT;
+    }
+}
