+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/sensor/LightSensorEventHandler.java
@@ -1,0 +1,32 @@
+package ru.yandex.practicum.telemetry.collector.handler.sensor;
+
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.LightSensorProto;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.LightSensorAvro;
+
+/**
+ * Обработчик события от датчика освещения
+ */
+@Component
+public class LightSensorEventHandler extends BaseSensorEventHandler<LightSensorAvro> {
+    public LightSensorEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    protected LightSensorAvro mapToAvro(SensorEventProto event) {
+        LightSensorProto lightSensorEvent = event.getLightSensorEvent();
+
+        return LightSensorAvro.newBuilder()
+                .setLinkQuality(lightSensorEvent.getLinkQuality())
+                .setLuminosity(lightSensorEvent.getLuminosity())
+                .build();
+    }
+
+    @Override
+    public SensorEventProto.PayloadCase getEventType() {
+        return SensorEventProto.PayloadCase.LIGHT_SENSOR_EVENT;
+    }
+}
