+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/sensor/MotionSensorEventHandler.java
@@ -1,0 +1,33 @@
+package ru.yandex.practicum.telemetry.collector.handler.sensor;
+
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.MotionSensorProto;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.MotionSensorAvro;
+
+/**
+ * Обработчик события от датчика движения
+ */
+@Component
+public class MotionSensorEventHandler extends BaseSensorEventHandler<MotionSensorAvro> {
+    public MotionSensorEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    protected MotionSensorAvro mapToAvro(SensorEventProto event) {
+        MotionSensorProto motionSensorEvent = event.getMotionSensorEvent();
+
+        return MotionSensorAvro.newBuilder()
+                .setMotion(motionSensorEvent.getMotion())
+                .setLinkQuality(motionSensorEvent.getLinkQuality())
+                .setVoltage(motionSensorEvent.getVoltage())
+                .build();
+    }
+
+    @Override
+    public SensorEventProto.PayloadCase getEventType() {
+        return SensorEventProto.PayloadCase.MOTION_SENSOR_EVENT;
+    }
+}
