+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/sensor/SwitchSensorEventHandler.java
@@ -1,0 +1,31 @@
+package ru.yandex.practicum.telemetry.collector.handler.sensor;
+
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.grpc.telemetry.event.SwitchSensorProto;
+import ru.yandex.practicum.kafka.telemetry.event.SwitchSensorAvro;
+
+/**
+ * Обработчик события от датчика переключателя
+ */
+@Component
+public class SwitchSensorEventHandler extends BaseSensorEventHandler<SwitchSensorAvro> {
+    public SwitchSensorEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    protected SwitchSensorAvro mapToAvro(SensorEventProto event) {
+        SwitchSensorProto switchSensorEvent = event.getSwitchSensorEvent();
+
+        return SwitchSensorAvro.newBuilder()
+                .setState(switchSensorEvent.getState())
+                .build();
+    }
+
+    @Override
+    public SensorEventProto.PayloadCase getEventType() {
+        return SensorEventProto.PayloadCase.SWITCH_SENSOR_EVENT;
+    }
+}
