+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/handler/sensor/TemperatureSensorEventHandler.java
@@ -1,0 +1,32 @@
+package ru.yandex.practicum.telemetry.collector.handler.sensor;
+
+import ru.yandex.practicum.telemetry.collector.kafka.Producer;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.grpc.telemetry.event.TemperatureSensorProto;
+import ru.yandex.practicum.kafka.telemetry.event.TemperatureSensorAvro;
+
+/**
+ * Обработчик события от температурного датчика
+ */
+@Component
+public class TemperatureSensorEventHandler extends BaseSensorEventHandler<TemperatureSensorAvro> {
+    public TemperatureSensorEventHandler(Producer producer) {
+        super(producer);
+    }
+
+    @Override
+    protected TemperatureSensorAvro mapToAvro(SensorEventProto event) {
+        TemperatureSensorProto temperatureSensorEvent = event.getTemperatureSensorEvent();
+
+        return TemperatureSensorAvro.newBuilder()
+                .setTemperatureC(temperatureSensorEvent.getTemperatureC())
+                .setTemperatureF(temperatureSensorEvent.getTemperatureF())
+                .build();
+    }
+
+    @Override
+    public SensorEventProto.PayloadCase getEventType() {
+        return SensorEventProto.PayloadCase.TEMPERATURE_SENSOR_EVENT;
+    }
+}
