+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/kafka/TopicType.java
@@ -1,0 +1,6 @@
+package ru.yandex.practicum.telemetry.collector.kafka;
+
+public enum TopicType {
+    SENSORS_EVENTS_TOPIC,
+    HUBS_EVENTS_TOPIC
+}
